# endereco-service/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8082'
basePath: /
tags:
- name: endereco-service
  description: Operações para gerenciamento de endereços
paths:
  /enderecos:
    get:
      tags:
      - endereco-controller
      summary: Listar todos os endereços
      operationId: buscarTodos
      produces:
      - '*/*'
      - application/json
      responses:
        '200':
          description: Dados recuperados com sucesso
          schema:
            type: array
            items:
              $ref: '#/definitions/endereco'
        '204':
          description: Não existem dados para serem recuperados
          schema:
            type: array
            items:
              $ref: '#/definitions/endereco'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
      - endereco-controller
      summary: Salva um novo endereço
      operationId: salvarEndereco
      consumes:
      - application/json
      produces:
      - '*/*'
      - application/json
      parameters:
      - in: body
        name: endereco
        description: endereco
        required: true
        schema:
          $ref: '#/definitions/endereco'
      responses:
        '200':
          description: OK
        '201':
          description: Operação realizada com sucesso
        '400':
          description: Dados invalidos
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /enderecos/{id}:
    get:
      tags:
      - endereco-controller
      summary: Recupera um endereco por id
      operationId: buscarPorId
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: string
      responses:
        '200':
          description: Operação realizada com sucesso
          schema:
            $ref: '#/definitions/endereco'
        '400':
          description: Dados invalidos
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Recurso não encontrado
    put:
      tags:
      - endereco-controller
      summary: Atualiza informações de um endereco por id
      operationId: autalizarEndereco
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: Id
        required: true
        type: string
      - in: body
        name: endereco
        description: endereco
        required: true
        schema:
          $ref: '#/definitions/endereco'
      responses:
        '200':
          description: Operação realizada com sucesso
          schema:
            $ref: '#/definitions/endereco'
        '201':
          description: Created
        '400':
          description: Dados invalidos
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Recurso não encontrado
    delete:
      tags:
      - endereco-controller
      summary: Remove um endereco por id
      operationId: deletarEndereco
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: string
      - in: body
        name: endereco
        description: endereco
        required: true
        schema:
          $ref: '#/definitions/endereco'
      responses:
        '200':
          description: Operação realizada com sucesso
        '204':
          description: No Content
        '400':
          description: Dados invalidos
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Recurso não encontrado
definitions:
  endereco:
    type: object
    properties:
      id:
        type: string
      rua:
        type: string
      complemento:
        type: string
      cep:
        type: string
      numero:
        type: number
      cidade:
        type: string
      estado:
        type: string
    title: endereco